+++ packages/apiclient/src/fetchApiConnector.ts
@@ -2,20 +2,24 @@ import { bind } from "@frui.ts/helpers";
 import FetchError from "./fetchError";
 import type { IApiConnector } from "./types";
 
-const jsonContentType = "application/json";
+export const jsonContentType = "application/json";
 export type Middleware = (response: Response) => Response | PromiseLike<Response>;
 export type FetchFunction = (input: RequestInfo, init?: RequestInit) => Promise<Response>;
-export type JsonSerializer = (value: any) => string;
+export type Serializer = (value: any) => string;
 
-/** Creates a new RequestInit based on the provided values and with a 'Content-Type: application/json' header. */
-export function appendJsonHeader(params?: RequestInit): RequestInit {
+export function appendContentTypeHeader(contentType: string, params?: RequestInit): RequestInit {
   return {
     ...params,
     headers: {
       ...(params ?? {}).headers,
-      "Content-Type": jsonContentType,
+      "Content-Type": contentType,
     },
   };
+}
+
+/** Creates a new RequestInit based on the provided values and with a 'Content-Type: application/json' header. */
+export function appendJsonHeader(params?: RequestInit): RequestInit {
+  return appendContentTypeHeader(jsonContentType, params);
 }
 
 /** Middleware used by FetchApiConnector to handle response status codes other than 2xx as errors */
@@ -42,12 +46,12 @@ export async function handleErrorStatusMiddleware(response: Response) {
  */
 export class FetchApiConnector implements IApiConnector {
   protected fetchFunction: FetchFunction;
-  protected jsonSerializer: JsonSerializer;
+  protected serializer: Serializer;
   protected middleware: Middleware;
 
-  constructor(configuration?: { fetchFunction?: FetchFunction; jsonSerializer?: JsonSerializer; middleware?: Middleware }) {
+  constructor(configuration?: { fetchFunction?: FetchFunction; serializer?: Serializer; middleware?: Middleware }) {
     this.fetchFunction = configuration?.fetchFunction ?? bind(window.fetch, window);
-    this.jsonSerializer = configuration?.jsonSerializer ?? JSON.stringify;
+    this.serializer = configuration?.serializer ?? JSON.stringify;
     this.middleware = configuration?.middleware ?? handleErrorStatusMiddleware;
   }
 
@@ -58,28 +62,28 @@ export class FetchApiConnector implements IApiConnector {
   post(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
     return this.fetchFunction(url, { ...params, method: "POST", body }).then(this.middleware);
   }
-  postJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.post(url, this.jsonSerializer(content), appendJsonHeader(params));
+  postObject(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.post(url, this.serializer(content), appendJsonHeader(params));
   }
 
   put(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
     return this.fetchFunction(url, { ...params, method: "PUT", body }).then(this.middleware);
   }
-  putJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.put(url, this.jsonSerializer(content), appendJsonHeader(params));
+  putObject(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.put(url, this.serializer(content), appendJsonHeader(params));
   }
 
   patch(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
     return this.fetchFunction(url, { ...params, method: "PATCH", body }).then(this.middleware);
   }
-  patchJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.patch(url, this.jsonSerializer(content), appendJsonHeader(params));
+  patchObject(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.patch(url, this.serializer(content), appendJsonHeader(params));
   }
 
   delete(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
     return this.fetchFunction(url, { ...params, method: "DELETE", body }).then(this.middleware);
   }
-  deleteJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.delete(url, this.jsonSerializer(content), appendJsonHeader(params));
+  deleteObject(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.delete(url, this.serializer(content), appendJsonHeader(params));
   }
 }