+++ packages/apiclient/src/restRequestBuilder.ts
@@ -1,8 +1,10 @@
-import type { StringifiableRecord, StringifyOptions } from "query-string";
-import { stringify } from "query-string";
+import type { IStringifyOptions } from "qs";
+import { stringify } from "qs";
 import type { IApiConnector } from "./types";
 
 const cleanupRegex = /\/+$/g; // removes trailing slash
+
+export type Deserializer<T> = (response: Response) => Promise<T>;
 
 export const ContentTypes = {
   json: "application/json,text/json",
@@ -23,12 +25,18 @@ export function appendUrl(base: string, ...segments: any[]) {
   return segments.length ? `${basePath}/${segments.join("/")}` : basePath;
 }
 
-/** Fluent URL builder that makes the network call with the underlying IApiConnector */
+/**
+ * Fluent URL builder that makes the network call with the underlying IApiConnector
+ * Check https://github.com/ljharb/qs for query string customizations
+ */
 export class RestRequestBuilder {
-  static DefaultQueryStringOptions: StringifyOptions = { skipNull: true };
+  static DefaultQueryStringOptions: IStringifyOptions = { skipNulls: true };
 
   protected urlValue: string;
-  queryStringOptions?: StringifyOptions;
+  queryStringOptions?: IStringifyOptions;
+
+  objectContentType: string = ContentTypes.json;
+  objectDeserializer: Deserializer<unknown> = x => x.json();
 
   get url() {
     return this.urlValue;
@@ -61,24 +69,24 @@ export class RestRequestBuilder {
     return this;
   }
 
-  get<T>(queryParams?: StringifiableRecord): Promise<T> {
+  get<T>(queryParams?: any): Promise<T> {
     const requestUrl = this.appendQuery(this.urlValue, queryParams);
-    const params = appendAcceptHeader(this.params, ContentTypes.json);
-    return this.apiConnector.get(requestUrl, params).then(x => x.json() as Promise<T>);
+    const params = appendAcceptHeader(this.params, this.objectContentType);
+    return this.apiConnector.get(requestUrl, params).then(this.objectDeserializer as Deserializer<T>);
   }
 
-  getRaw(queryParams?: StringifiableRecord) {
+  getRaw(queryParams?: any) {
     const requestUrl = this.appendQuery(this.urlValue, queryParams);
     return this.apiConnector.get(requestUrl, this.params);
   }
 
   post<T>(content: any): Promise<T> {
-    const params = appendAcceptHeader(this.params, ContentTypes.json);
-    return this.apiConnector.postJson(this.urlValue, content, params).then(x => x.json() as Promise<T>);
+    const params = appendAcceptHeader(this.params, this.objectContentType);
+    return this.apiConnector.postObject(this.urlValue, content, params).then(this.objectDeserializer as Deserializer<T>);
   }
 
   postOnly(content: any) {
-    return this.apiConnector.postJson(this.urlValue, content, this.params);
+    return this.apiConnector.postObject(this.urlValue, content, this.params);
   }
 
   postData(data?: BodyInit) {
@@ -86,12 +94,12 @@ export class RestRequestBuilder {
   }
 
   put<T>(content: any): Promise<T> {
-    const params = appendAcceptHeader(this.params, ContentTypes.json);
-    return this.apiConnector.putJson(this.urlValue, content, params).then(x => x.json() as Promise<T>);
+    const params = appendAcceptHeader(this.params, this.objectContentType);
+    return this.apiConnector.putObject(this.urlValue, content, params).then(this.objectDeserializer as Deserializer<T>);
   }
 
   putOnly(content: any) {
-    return this.apiConnector.putJson(this.urlValue, content, this.params);
+    return this.apiConnector.putObject(this.urlValue, content, this.params);
   }
 
   putData(data?: BodyInit) {
@@ -99,12 +107,12 @@ export class RestRequestBuilder {
   }
 
   patch<T>(content: any): Promise<T> {
-    const params = appendAcceptHeader(this.params, ContentTypes.json);
-    return this.apiConnector.patchJson(this.urlValue, content, params).then(x => x.json() as Promise<T>);
+    const params = appendAcceptHeader(this.params, this.objectContentType);
+    return this.apiConnector.patchObject(this.urlValue, content, params).then(this.objectDeserializer as Deserializer<T>);
   }
 
   patchOnly(content: any) {
-    return this.apiConnector.patchJson(this.urlValue, content, this.params);
+    return this.apiConnector.patchObject(this.urlValue, content, this.params);
   }
 
   patchData(data?: BodyInit) {
@@ -113,7 +121,7 @@ export class RestRequestBuilder {
 
   delete(content?: any) {
     return content
-      ? this.apiConnector.deleteJson(this.urlValue, content, this.params)
+      ? this.apiConnector.deleteObject(this.urlValue, content, this.params)
       : this.apiConnector.delete(this.urlValue, undefined, this.params);
   }
 
@@ -122,11 +130,11 @@ export class RestRequestBuilder {
     return this;
   }
 
-  getQueryString(query: StringifiableRecord, queryStringOptions?: StringifyOptions) {
+  getQueryString(query: any, queryStringOptions?: IStringifyOptions) {
     return stringify(query, queryStringOptions ?? this.queryStringOptions ?? RestRequestBuilder.DefaultQueryStringOptions);
   }
 
-  appendQuery(url: string, query?: StringifiableRecord, queryStringOptions?: StringifyOptions) {
+  appendQuery(url: string, query?: any, queryStringOptions?: IStringifyOptions) {
     if (!query) {
       return url;
     }